Library is able to send crashes upstream
The library now exposes an callback for uploading data to a server, and also provides API for manually uploading crashes. This means that the LibraryState has also a base url (which is not configurable via API).

Added new API for discarding available LOGs, and crashes are no longer dumped on the logcat (just saved locally).

Initialization of the CrashHandlerLibrary has been moved to `onCreate` (WTF? the constructor? That was a fial). Now uploading is being done from the main app, and not from an activity. The main app also defines a network uploader using OKHTTP4 as a backend.

TODO:
- make a default network uploade using stock Android API
- schedule a timer to periodically send upstream bugs